Split gadget4 tests.
