+++ tests/frontends/test_gadget4_1.py
@@ -10,10 +10,3 @@ class Gadget4ArborSingleTest(TempDirTest, ArborTest):
     groups = ("forest", "tree", "prog")
     num_data_files = 1
     tree_skip = 100
-
-class Gadget4ArborMultipleTest(TempDirTest, ArborTest):
-    arbor_type = Gadget4Arbor
-    test_filename = "gadget4/treedata/trees.0.hdf5"
-    groups = ("forest", "tree", "prog")
-    num_data_files = 64
-    tree_skip = 100
